Done with all the functionalities. Integrated Swagger, added the logger statements, used .env variables.

+++ src/main/java/com/mukesh/restaurantManagementSystem/dto/request/OrderItemsRequestDTO.java
@@ -18,5 +18,5 @@ public class OrderItemsRequestDTO {
     @NotNull
     Map<Long, Integer> foodItems;
     @NotNull
-    private Long tableNumber;
+    private Integer tableNumber;
 }
